+++ HvZ-TheGame/HvZ-TheGame/BaseClass.h
@@ -15,7 +15,7 @@ public:
 	BaseClass(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, sf::Vector3f pointLocation, int itemGroup);
 	~BaseClass();
 
-	virtual void move(int x, int y, int z);
+	virtual void move(float x, float y, float z);
 	virtual void hitDetect(BaseClass& test);
 	virtual void draw(sf::RenderWindow& window);
 };
